FIX API

+++ src/services/cartItem.service.js
@@ -3,6 +3,7 @@ const userService = require("../services/user.service.js");
 
 
 async function updateCartItem(userId,cartItemId,cartItemData){
+    console.log(userId,cartItemId,cartItemData)
     try{
         const item = await findCartItemById(cartItemId);
         if(!item){
@@ -12,6 +13,7 @@ async function updateCartItem(userId,cartItemId,cartItemData){
         if(!user){
             throw new Error("Không tìm thấy người dùng : ",userId)
         }
+        console.log("item", item.product)
         if(user._id.toString()===userId.toString()){
             item.quantity = cartItemData.quantity;
             item.price = item.quantity*item.product.price;
@@ -26,18 +28,27 @@ async function updateCartItem(userId,cartItemId,cartItemData){
         throw new Error(error.message)
     }
 }
-async function removeCartItem(userId,cartItemId){
-    const cartItem = await findCartItemById(cartItemId);
-    const user = await userService.findUserByEmail(userId);
-
-    if(user._id.toString()===cartItem.userId.toString()){
-        await CartItem.findByIdAndDelete(cartItemId)
-
+async function removeCartItem(userId, cartItemId) {
+    try {
+        const cartItem = await findCartItemById(cartItemId);
+        if (!cartItem) {
+            throw new Error("Không tìm thấy mục giỏ hàng");
+        }
+        const user = await userService.getUserById(userId);
+        if (!user) {
+            throw new Error("Người dùng không tồn tại");
+        }
+        if (user._id.toString() === cartItem.userId.toString()) {
+            return await CartItem.findByIdAndDelete(cartItemId);
+        }
+        throw new Error("Bạn không thể xóa mục của người dùng khác");
+    } catch (error) {
+        console.error(error.message);
+        throw error;
     }
-    throw new Error("bạn không thể xóa mục của người dùng khác");
 }
 async function findCartItemById(cartItemId){
-    const cartItem = await findCartItemById(cartItemId);
+    const cartItem = await CartItem.findById(cartItemId).populate("product");
     if(cartItem){
         return cartItem
